added upto Exercise27

+++ src/main/java/org/comit/practise/_01_variables/Exercise03.java
@@ -5,8 +5,8 @@ public class Exercise03 {
 	public static void main(String[] args) {
 		//Write a program in Java that prints a float number, for example, the 5.3 or the 7.5.
 		
-		double num1 = 5.3;
-		double num2 = 7.5;
+		float num1 = 5.3f;
+		float num2 = 7.5f;
 		
 		System.out.println("The first number is :" + num1);
 		System.out.println("The second number is :" + num2);
